average color
