+++ packages/convert/app/components/Probe.tsx
@@ -1,14 +1,4 @@
-import type {
-	Dimensions,
-	MediaParserAudioCodec,
-	MediaParserVideoCodec,
-	ParseMediaContainer,
-	TracksField,
-} from '@remotion/media-parser';
-import {parseMedia} from '@remotion/media-parser';
-import {fetchReader} from '@remotion/media-parser/fetch';
-import {webFileReader} from '@remotion/media-parser/web-file';
-import React, {useCallback, useEffect, useMemo, useState} from 'react';
+import React, {useCallback, useMemo, useState} from 'react';
 import {Source} from '~/lib/convert-state';
 import {AudioTrackOverview} from './AudioTrackOverview';
 import {ContainerOverview} from './ContainerOverview';
@@ -27,188 +17,25 @@ import {
 import {ScrollArea} from './ui/scroll-area';
 import {Separator} from './ui/separator';
 import {Skeleton} from './ui/skeleton';
+import {useProbe} from './use-probe';
 
 export const Probe: React.FC<{
 	readonly src: Source;
 	readonly setProbeDetails: React.Dispatch<React.SetStateAction<boolean>>;
 	readonly probeDetails: boolean;
 }> = ({src, probeDetails, setProbeDetails}) => {
-	const [audioCodec, setAudioCodec] = useState<MediaParserAudioCodec | null>(
-		null,
-	);
-	const [fps, setFps] = useState<number | null | undefined>(undefined);
-	const [durationInSeconds, setDurationInSeconds] = useState<number | null>(
-		null,
-	);
-	const [dimensions, setDimensions] = useState<Dimensions | null>(null);
-	const [name, setName] = useState<string | null>(null);
-	const [videoCodec, setVideoCodec] = useState<MediaParserVideoCodec | null>(
-		null,
-	);
-	const [size, setSize] = useState<number | null>(null);
-	const [tracks, setTracks] = useState<TracksField | null>(null);
-	const [container, setContainer] = useState<ParseMediaContainer | null>(null);
-	const [thumbnail, setThumbnail] = useState<VideoFrame | null>(null);
-
-	const getStart = useCallback(() => {
-		const controller = new AbortController();
-		let hasFps = false;
-		let hasDuration = false;
-		let hasDimensions = false;
-		let hasVideoCodec = false;
-		let hasAudioCodec = false;
-		let hasSize = false;
-		let hasName = false;
-		let hasFrame = false;
-		let hasContainer = false;
-		let hasTracks = false;
-
-		const cancelIfDone = () => {
-			if (
-				hasFps &&
-				hasDuration &&
-				hasDimensions &&
-				hasVideoCodec &&
-				hasAudioCodec &&
-				hasSize &&
-				hasName &&
-				hasFrame &&
-				hasContainer &&
-				hasTracks
-			) {
-				controller.abort(new Error('Cancelled (all info)'));
-			}
-		};
-
-		parseMedia({
-			src: src.type === 'file' ? src.file : src.url,
-			fields: {
-				dimensions: true,
-				videoCodec: true,
-				size: true,
-				durationInSeconds: true,
-				audioCodec: true,
-				fps: true,
-				name: true,
-				tracks: true,
-				container: true,
-			},
-			reader: src.type === 'file' ? webFileReader : fetchReader,
-			signal: controller.signal,
-			onVideoTrack: async (track) => {
-				if (typeof VideoDecoder === 'undefined') {
-					return null;
-				}
-
-				let frames = 0;
-
-				const decoder = new VideoDecoder({
-					error: (error) => {
-						// eslint-disable-next-line no-console
-						console.log(error);
-					},
-					output(frame) {
-						frames++;
-						if (frames < 30) {
-							frame.close();
-							return;
-						}
-
-						setThumbnail((f) => {
-							if (f) {
-								frame.close();
-								return f;
-							}
-
-							return frame;
-						});
-						hasFrame = true;
-						cancelIfDone();
-					},
-				});
-
-				if (!(await VideoDecoder.isConfigSupported(track)).supported) {
-					return null;
-				}
-
-				// TODO: See if possible
-				decoder.configure(track);
-				return (sample) => {
-					if (hasFrame) {
-						return;
-					}
-
-					decoder.decode(new EncodedVideoChunk(sample));
-				};
-			},
-			onContainer(c) {
-				hasContainer = true;
-				setContainer(c);
-				cancelIfDone();
-			},
-			onAudioCodec: (codec) => {
-				hasAudioCodec = true;
-				setAudioCodec(codec);
-				cancelIfDone();
-			},
-			onFps: (f) => {
-				hasFps = true;
-				setFps(f);
-				cancelIfDone();
-			},
-			onDurationInSeconds: (d) => {
-				hasDuration = true;
-				setDurationInSeconds(d);
-				cancelIfDone();
-			},
-			onName: (n) => {
-				hasName = true;
-				setName(n);
-				cancelIfDone();
-			},
-			onDimensions(dim) {
-				hasDimensions = true;
-				setDimensions(dim);
-				cancelIfDone();
-			},
-			onVideoCodec: (codec) => {
-				hasVideoCodec = true;
-				setVideoCodec(codec);
-				cancelIfDone();
-			},
-			onTracks: (trx) => {
-				hasTracks = true;
-				setTracks(trx);
-				cancelIfDone();
-			},
-			onSize(s) {
-				hasSize = true;
-				setSize(s);
-				cancelIfDone();
-			},
-		})
-			.then(() => {})
-			.catch((err) => {
-				if ((err as Error).stack?.includes('Cancelled')) {
-					return;
-				}
-				if ((err as Error).stack?.toLowerCase()?.includes('aborted')) {
-					return;
-				}
-
-				// eslint-disable-next-line no-console
-				console.log(err);
-			});
-
-		return controller;
-	}, [src]);
-
-	useEffect(() => {
-		const start = getStart();
-		return () => {
-			start.abort(new Error('Cancelled (strict mode)'));
-		};
-	}, [getStart]);
+	const {
+		audioCodec,
+		fps,
+		tracks,
+		name,
+		thumbnail,
+		container,
+		dimensions,
+		size,
+		videoCodec,
+		durationInSeconds,
+	} = useProbe(src);
 
 	const onClick = useCallback(() => {
 		setProbeDetails((p) => !p);
