+++ packages/convert/package.json
@@ -30,7 +30,8 @@
 		"react": "18.3.1",
 		"react-dom": "18.3.1",
 		"tailwind-merge": "2.5.2",
-		"tailwindcss-animate": "^1.0.7"
+		"tailwindcss-animate": "^1.0.7",
+		"fast-average-color": "9.4.0"
 	},
 	"devDependencies": {
 		"@remix-run/dev": "^2.11.2",
